dom/html/test/test_bug2082.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=2082
-->
<head>
  <title>Test for Bug 2082</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>        
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=2082">Mozilla Bug 2082</a>
<p id="display"></p>
<div id="content" style="display: none">
<FORM name="gui" id="gui">
<INPUT TYPE="text" NAME="field" VALUE="some value">
</FORM>  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 2082 **/
var guiform = document.getElementById("gui");
ok(document.getElementById("gui").hasChildNodes(), "form elements should be treated as form's children");

</script>
</pre>
</body>
</html>
